# documentation/api-documentation.yml
openapi: "3.0.0"
info:
  version: "3.0.1"
  title: "FemDevs API Documentation"
  description: |-
    The documentation for the FemDevs API.
    Do note that some routes have permission-based requirements that your user must be quilified for in order to access them.
    Check the description for each route in case you are unsure.

    Permissions Overrides:

    The following permissions will bypass all permissions unless its marked otherwise:
    * `Global::Role.Developer`
    * `Global::Role.Administrator`
    * `Global::Role.Owner`
  termsOfService: "https://legal.thefemdevs.com/terms"
  contact:
    name: "FemDevs"
    url: "https://thefemdevs.com"
    email: "support@thefemdevs.com"
  license:
    name: "GNU Affero General Public License v3.0"
    url: "https://www.gnu.org/licenses/agpl-3.0.en.html"
servers:
  - url: "https://api.thefemdevs.com/"
    description: "Production Server"
paths:
  /user/get:
    get:
      operationId: "GetUser"
      description: |-
        Gets the user information for the user with the given ID.
        Returns the user information if the user exists, otherwise returns an error.

        Permissions Required: `Global::User.Read` unless requesting own information"
      summary: "Get User Info by ID"
      parameters:
        - in: "header"
          name: "x-uid"
          description: "The Firebase UID of the user to get information for"
          schema:
            type: string
      responses:
        200:
          description: "The user information for the given ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetUserInfoResponse"
              examples:
                user:
                  summary: "Example User"
                  value:
                    id: "gPcqw6MAbXbZs6ARohHqT2vapFY2"
                    username: "JohnDoe0001"
                    fullName: "John Doe"
                    email: "example@example.com"
                    permissions: 0
        400:
          description: "Missing user information"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-User:
                  summary: "Missing UID Header"
                  value:
                    code: 8
                    message: "Missing Header Parameter(s)"
                Invalid-User:
                  summary: "Invalid UID Header"
                  value:
                    code: 9
                    message: "Invalid Header Parameter(s)"
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
        404:
          description: "Error with the given ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 22
                    message: "Unable to retreive requested information"
      security:
        - APIToken: []
  /user/update:
    patch:
      operationId: "UpdateUser"
      description: |-
        Updates the user information for the user with the given ID.
        Returns the user information if the user exists, otherwise returns an error.

        Permissions Required: `Global::User.Write` unless updating own user information with few field exceptions
      summary: "Update User Info by ID"
      parameters:
        - in: "header"
          name: "x-uid"
          description: "The Firebase UID of the user to get information for"
          schema:
            type: string
      requestBody:
        description: "The new user information"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInfo"
            examples:
              displayname:
                value:
                  displayname: "newDisplayName"
              email:
                value:
                  email: "newEmail@example.com"
              password:
                value:
                  password: "newPassword"
              perms:
                value:
                  perms: 1
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateUserInfoResponse"
              examples:
                displayname:
                  summary: "Update Response for Display Name"
                  value:
                    id: "gPcqw6MAbXbZs6ARohHqT2vapFY2"
                    updatedUser:
                      id: 19
                      firebaseuid: "gPcqw6MAbXbZs6ARohHqT2vapFY2"
                      firstname: "John"
                      lastname: "Doe"
                      displayname: "newDisplayName"
                      email: "example@example.com"
                      dateofbirth: "2000-01-01"
                      permissions: 0
                      betaAccess: 0
                email:
                  summary: "Update Response for Email"
                  value:
                    id: "gPcqw6MAbXbZs6ARohHqT2vapFY2"
                    updatedUser:
                      id: 19
                      firebaseuid: "gPcqw6MAbXbZs6ARohHqT2vapFY2"
                      firstname: "John"
                      lastname: "Doe"
                      displayname: "JohnDoe0001"
                      email: "newEmail@example.com"
                      dateofbirth: "2000-01-01"
                      permissions: 0
                      betaAccess: 0
                password:
                  summary: "Update Response for Password"
                  value:
                    id: "gPcqw6MAbXbZs6ARohHqT2vapFY2"
                    updatedUser:
                      id: 19
                      firebaseuid: "gPcqw6MAbXbZs6ARohHqT2vapFY2"
                      firstname: "John"
                      lastname: "Doe"
                      displayname: "JohnDoe0001"
                      email: "example@example.com"
                      dateofbirth: "2000-01-01"
                      permissions: 0
                      betaAccess: 0
                perms:
                  summary: "Update Response for Permissions"
                  value:
                    id: "gPcqw6MAbXbZs6ARohHqT2vapFY2"
                    updatedUser:
                      id: 19
                      firebaseuid: "gPcqw6MAbXbZs6ARohHqT2vapFY2"
                      firstname: "John"
                      lastname: "Doe"
                      displayname: "JohnDoe0001"
                      email: "example@example.com"
                      dateofbirth: "2000-01-01"
                      permissions: 0
                      betaAccess: 1
        400:
          description: "Missing user information"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-User:
                  summary: "Missing UID Header"
                  value:
                    code: 8
                    message: "Missing Header Parameter(s)"
                Invalid-User:
                  summary: "Invalid UID Header"
                  value:
                    code: 9
                    message: "Invalid Header Parameter(s)"
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
        404:
          description: "Error with the given ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 22
                    message: "Unable to retreive requested resource"
        500:
          description: "Unable to update user"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 16
                    message: "Unable to update user"
      security:
        - APIToken: []
  /user/create:
    post:
      operationId: "CreateUser"
      description: |-
        Creates a user using the provided information.
        Returns the user information if successful, otherwise returns an error.

        Permissions Required: `Global::User.Create`
      summary: "Create a User"
      requestBody:
        description: "The new user information"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserInfo"
            examples:
              user:
                value:
                  username: "JohnDoe0001"
                  firstname: "John"
                  lastname: "Doe"
                  email: "example@example.com"
                  password: "ARandomPassword"
      responses:
        201:
          description: "The user information for the given ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserInfoResponse"
              examples:
                user:
                  summary: "Example User"
                  value:
                    user:
                      uid: "gPcqw6MAbXbZs6ARohHqT2vapFY2"
                      email: "example@example.com"
                      emailVerified: false
                      displayName: "JohnDoe0001"
                      disabled: false
                      metadata:
                        lastSignInTime: "null"
                        creationTime: "Sat, 23 Sep 2023 16:06:10 GMT"
                        lastRefreshTime: "null"
                      tokensValidAfterTime: "Sat, 23 Sep 2023 16:06:10 GMT"
                      providerData:
                        - uid: "example@example.com"
                          displayName: "JohnDoe0001"
                          email: "example@example.com"
                          providerId: "password"
                    username: "JohnDoe0001"
                    fullName: "John Doe"
                    email: "example@example.com"
                    permissions: 0
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
        404:
          description: "Error with the given ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 22
                    message: "Unable to retreive requested information"
        500:
          description: "Unable to create user"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 15
                    message: "Unable to create user"
      security:
        - APIToken: []
  /user/delete:
    delete:
      operationId: "DeleteUser"
      description: |-
        Deletes the user with the given ID.
        Returns `true if successful, otherwise returns an error.

        Permissions Required: `Global::User.Delete`
      summary: "Delete User by ID"
      parameters:
        - in: "header"
          name: "x-uid"
          description: "The Firebase UID of the user to get delete"
          schema:
            type: string
      responses:
        200:
          description: "The successful notification of the deletion"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteUserInfoResponse"
              examples:
                deleted:
                  summary: "User Deleted"
                  value:
                    uid: "gPcqw6MAbXbZs6ARohHqT2vapFY2"
                    deleted: true
        400:
          description: "Missing user information"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-User:
                  summary: "Missing UID Header"
                  value:
                    code: 8
                    message: "Missing Header Parameter(s)"
                Invalid-User:
                  summary: "Invalid UID Header"
                  value:
                    code: 9
                    message: "Invalid Header Parameter(s)"
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
        404:
          description: "Error with the given ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 22
                    message: "Unable to retreive requested resource"
        500:
          description: "Error with the given ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 17
                    message: "Unable to delete requested user"
      security:
        - APIToken: []
...
components:
  securitySchemes:
    APIToken:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: |-
        JSON Web Token Authentication with information containing your license key

        Permissions:
        - `Global::Data.Read` (0:0:0) - **Unused**
        - `Global::Data.Write` (0:0:1) - **Unused**
        - `Global::Data.Create` (0:0:2) - **Unused**
        - `Global::Data.Delete` (0:0:3) - **Unused**
        - `Global::Token.Read` (0:1:0) - Allows reading of token information
        - `Global::Token.ReadAll` (0:1:1) - Allows reading of all token information
        - `Global::Token.Write` (0:1:2) - **Unused**
        - `Global::Token.Create` (0:1:3) - Allows creating of token information
        - `Global::Token.Delete` (0:1:4) - Allows deleting of token 
        - `Global::User.Read` (0:2:0) - Read user information
        - `Global::User.ReadAll` (0:2:1) - **Unused**
        - `Global::User.Write` (0:2:2) - Write/Update user information
        - `Global::User.Create` (0:2:3) - Create a user profile
        - `Global::User.Delete` (0:2:4) - Delete a user profile
        - `Cryptography::Encrypt` (1:0) - Ability to encrypt data using the `/crypto/enc` endpoint
        - `Cryptography::Decrypt` (1:1) - Ability to decrypt data using the `/crypto/dec` endpoint
        - `Dictionary::Definition` (2:0) - Ability to get the definition of a word using the `/dict/def` endpoint
        - `Dictionary::Synonym` (2:1) - Ability to get the synonyms of a word using the `/dict/syn` endpoint
        - `Dictionary::Antonym` (2:2) - Ability to get the antonyms of a word using the `/dict/ant` endpoint
        - `Facts::Dog` (3:0) - Ability to get a random dog fact using the `/facts/dog` endpoint
        - `Facts::Cat` (3:1) - Ability to get a random cat fact using the `/facts/cat` endpoint
        - `Facts::Random` (3:2) - Ability to get a random fact using the `/facts/random` endpoint
        - `Location::Coord` (4:0) - Ability to get the location of coordinates using the `/location/coords` endpoint
        - `Location::Pluscode` (4:1) - Ability to get the location of a pluscode using the `/location/pluscode` endpoint
        - `Location::Address` (4:2) - Ability to get the location of an address using the `/location/address` endpoint
        - `Phone::Lookup` (5:0) - Ability to get the details of a phone number using the `/phone/details` endpoint
        - `Phone::Validate` (5:1) - **Unused**
        - `Quote::Kanye` (6:0) - Ability to get a random Kanye West quote using the `/quotes/kanye` endpoint
        - `Quote::RonSwanson` (6:1) - Ability to get a random Ron Swanson quote using the `/quotes/ronswanson` endpoint
        - `Quote::Random` (6:2) - Ability to get a random quote using the `/quotes/random` endpoint
        - `Weather::Current` (7:0) - Ability to get the current weather using the `/weather/current` endpoint
        - `Minecraft::Hive.Player` (8:0:0) - Ability to get information about a player on the Hive Minecraft server using the `/minecraft/hive/player` endpoint
...
  schemas:
    FirebaseProviderData:
      description: |-
        The Authentication Provider Data for a Firebase User
      required:
        - uid
        - displayName
        - email
        - providerId
      properties:
        uid:
          nullable: false
          type: string
          description: "The Firebase UID of the user"
        displayName:
          nullable: false
          type: string
          description: "The display name of the user"
        email:
          nullable: false
          type: string
          description: "The email of the user"
        providerId:
          nullable: false
          type: string
          description: "The provider ID of the user"
    FirebaseMetadata:
      description: |-
        The Metadata for a Firebase User
      required:
        - lastSignInTime
        - creationTime
        - lastRefreshTime
      properties:
        lastSignInTime:
          nullable: false
          type: string
          description: "The last sign in time of the user"
        creationTime:
          nullable: false
          type: string
          description: "The creation time of the user"
        lastRefreshTime:
          nullable: false
          type: string
          description: "The last refresh time of the user"
    CreatedFirebaseUser:
      description: |-
        The Created Firebase User Information
      required:
        - uid
        - displayName
        - email
        - emailVerified
        - phoneNumber
        - photoURL
        - disabled
        - metadata
        - providerData
      properties:
        uid:
          nullable: false
          type: string
          description: "The Firebase UID of the user"
        displayName:
          nullable: false
          type: string
          description: "The display name of the user"
        email:
          nullable: false
          type: string
          description: "The email of the user"
        emailVerified:
          nullable: false
          type: boolean
          description: "Whether or not the email is verified"
        phoneNumber:
          nullable: false
          type: string
          description: "The phone number of the user"
        photoURL:
          nullable: false
          type: string
          description: "The photo URL of the user"
        disabled:
          nullable: false
          type: boolean
          description: "Whether or not the user is disabled"
        metadata:
          nullable: false
          type: object
          $ref: "#/components/schemas/FirebaseMetadata"
          description: "The metadata of the user"
        providerData:
          nullable: false
          type: array
          items:
            $ref: "#/components/schemas/FirebaseProviderData"
          description: "The provider data of the user"
    FullUserInformation:
      description: |-
        The Full User Information
      required:
        - id
        - firebaseuid
        - firstname
        - lastname
        - displayname
        - email
        - dateofbirth
        - permissions
        - betaAccess
      properties:
        id:
          nullable: false
          type: integer
          description: "The ID of the user"
        firebaseuid:
          nullable: false
          type: string
          description: "The Firebase UID of the user"
        firstname:
          nullable: false
          type: string
          description: "The first name of the user"
        lastname:
          nullable: false
          type: string
          description: "The last name of the user"
        displayname:
          nullable: false
          type: string
          description: "The display name of the user"
        email:
          nullable: false
          type: string
          description: "The email of the user"
        dateofbirth:
          nullable: false
          type: string
          description: "The date of birth of the user"
        permissions:
          nullable: false
          type: integer
          description: "The permissions of the user"
        betaAccess:
          nullable: false
          type: integer
          description: "Whether or not the user has beta access"
    GetUserInfoResponse:
      description: |-
        The information that is returned when getting user information
      required:
        - id
        - username
        - fullName
        - email
        - permissions
      properties:
        id:
          nullable: false
          type: string
          description: "The Firebase UID of the user"
        username:
          nullable: false
          type: string
          description: "The username of the user"
        fullName:
          nullable: false
          type: string
          description: "The full name of the user"
        email:
          nullable: false
          type: string
          description: "The email of the user"
        permissions:
          nullable: false
          type: integer
          description: "The permissions of the user"
    UpdateUserInfo:
      description: |-
        The information that is used to update user information
      properties:
        displayname:
          nullable: true
          type: string
          description: "The new display name of the user"
        email:
          nullable: true
          type: string
          description: "The new email of the user"
        password:
          nullable: true
          type: string
          description: "The new password of the user"
        perms:
          nullable: true
          type: integer
          description: "The new permissions of the user"
    UpdateUserInfoResponse:
      description: |-
        The information that is returned when updating user information
      required:
        - id
        - upatedUser
      properties:
        id:
          nullable: false
          type: string
          description: "The Firebase UID of the user"
        updatedUser:
          nullable: false
          type: object
          $ref: "#/components/schemas/FullUserInformation"
          description: "The updated user information"
    CreateUserInfo:
      description: |-
        The information that is used and required to create a new user
      required:
        - username
        - firstname
        - lastname
        - email
        - password
      properties:
        username:
          nullable: false
          type: string
          description: "The username of the user"
        firstname:
          nullable: false
          type: string
          description: "The first name of the user"
        lastname:
          nullable: false
          type: string
          description: "The last name of the user"
        email:
          nullable: false
          type: string
          description: "The email of the user"
        password:
          nullable: false
          type: string
          description: "The password of the user"
    CreateUserInfoResponse:
      description: |-
        The information that is returned when creating a new user
      required:
        - user
        - username
        - fullName
        - email
        - permissions
      properties:
        user:
          nullable: false
          type: object
          $ref: "#/components/schemas/CreatedFirebaseUser"
          description: "The user information"
        username:
          nullable: false
          type: string
          description: "The username of the user"
        fullName:
          nullable: false
          type: string
          description: "The full name of the user"
        email:
          nullable: false
          type: string
          description: "The email of the user"
        permissions:
          nullable: false
          type: integer
          description: "The permissions of the user"
    DeleteUserInfoResponse:
      description: |-
        The information that is returned when deleting a user
      required:
        - uid
        - deleted
      properties:
        uid:
          nullable: false
          type: string
          description: "The Firebase UID of the user"
        deleted:
          nullable: false
          type: boolean
          description: "Whether or not the user was deleted"
    ErrorData:
      description: |-
        The error data that is returned when an error occurs

        All Error Codes are located in `/functions/errorResponse.js` [GitHub](https://github.com/femdevs/femdev-website/blob/main/functions/errorResponse.js)
      required:
        - code
        - message
      properties:
        code:
          nullable: false
          type: integer
          description: "The error code"
        message:
          nullable: false
          type: string
          description: "The error message"
...